third_party/blink/web_tests/http/tests/devtools/elements/elements-panel-limited-children-expected.txt

Tests that src and href element targets are rewritten properly.

=========== Loaded 5 children ===========
- <html>
    + <head>…</head>
    - <body>
        - <div id="data">
              <div id="id1">1</div>
              <div id="id2">2</div>
              <div id="id3">3</div>
              <div id="id4">4</div>
              <div id="id5">5</div>
              Show All Nodes (5 More)
          </div>
      </body>
  </html>
=========== Modified children ===========
- <html>
    + <head>…</head>
    - <body>
        - <div id="data">
              <a></a>
              <div id="id1">1</div>
              <div id="id3">3</div>
              <div id="id4">4</div>
              <div id="id5">5</div>
              Show All Nodes (6 More)
          </div>
      </body>
  </html>
=========== Loaded all children ===========
- <html>
    + <head>…</head>
    - <body>
        - <div id="data">
              <a></a>
              <div id="id1">1</div>
              <div id="id3">3</div>
              <div id="id4">4</div>
              <div id="id5">5</div>
              <div id="id6">6</div>
              <div id="id7">7</div>
              <div id="id8">8</div>
              <div id="id9">9</div>
              <div id="id10">10</div>
              <a></a>
          </div>
      </body>
  </html>
